requirements.txt
cloudshell-core==2.2.176
cloudshell-l1-networking-core>=1.0,<1.1
cloudshell-snmp>=3.1.0,<3.2.0